3) Développer la route POST /games : **/!\ La console du jeu ne sera pas créée dans cette route. Il faut que la console existe déjà.**
    4) Développer la route PATCH /games/{id}
4)
    1) Développer le modèle Review
    2) Développer le DTO Review
    3) Développer la route GET /reviews
    4) Développer la route GET /reviews/{id}
    5) Développer la route POST /reviews : **/!\ Le jeu de la review ne sera pas créé dans cette route. Il faut que le jeu existe déjà.**
    6) Développer la route PATCH /reviews/{id}

Début question 5.1

+++ src/controllers/console.controller.ts
@@ -1,5 +1,6 @@
 import { Controller, Get, Post, Delete, Route, Path, Body, Tags, Patch } from "tsoa";
 import { consoleService } from "../services/console.service";
+import { reviewService } from "../services/review.service"; // Importer le service des reviews
 import { ConsoleDTO } from "../dto/console.dto";
 
 @Route("consoles")
@@ -29,7 +30,13 @@ export class ConsoleController extends Controller {
   // Supprime une console par ID
   @Delete("{id}")
   public async deleteConsole(@Path() id: number): Promise<void> {
+    const hasReviews = await reviewService.hasReviewsForConsole(id);
+    if (hasReviews) {
+      this.setStatus(400); // Bad Request
+      throw new Error("Cannot delete console with existing reviews for its games.");
+    }
     await consoleService.deleteConsole(id);
+    this.setStatus(204); // No Content
   }
 
   // Met à jour une console par ID
@@ -40,6 +47,5 @@ export class ConsoleController extends Controller {
   ): Promise<ConsoleDTO | null> {
     const { name, manufacturer } = requestBody;
     return await consoleService.updateConsole(id, name, manufacturer);
-    
   }
 }
